moved SocialRepository to shared module and registered it inside sharedRepositoryModule
