+++ shared/src/commonMain/kotlin/org/listenbrainz/shared/di/SharedRepositoryModule.kt
@@ -21,6 +21,8 @@ import org.listenbrainz.shared.repository.brainzplayer.SongRepository
 import org.listenbrainz.shared.repository.brainzplayer.SongRepositoryImpl
 import org.listenbrainz.shared.repository.brainzplayer.PlaylistRepository
 import org.listenbrainz.shared.repository.brainzplayer.PlaylistRepositoryImpl
+import org.listenbrainz.shared.repository.social.SocialRepository
+import org.listenbrainz.shared.repository.social.SocialRepositoryImpl
 
 val sharedRepositoryModule = module {
     single<SocketRepository> { SocketRepositoryImpl(get<HttpClient>(named(SHARED_HTTP_CLIENT)),get()) }
@@ -41,4 +43,5 @@ val sharedRepositoryModule = module {
     }
     single<SongRepository> { SongRepositoryImpl(get(),get()) }
     single<PlaylistRepository> { PlaylistRepositoryImpl(get()) }
+    single<SocialRepository> { SocialRepositoryImpl(get(),get()) }
 }
