+++ shared/src/commonMain/kotlin/org/listenbrainz/shared/repository/social/SocialRepositoryImpl.kt
@@ -1,20 +1,19 @@
-package org.listenbrainz.android.repository.social
+package org.listenbrainz.shared.repository.social
 
-import org.listenbrainz.android.model.PinData
+import org.listenbrainz.shared.model.PinData
 import org.listenbrainz.shared.model.PinnedRecording
-import org.listenbrainz.android.model.RecommendationData
+import org.listenbrainz.shared.model.RecommendationData
 import org.listenbrainz.shared.model.ResponseError
-import org.listenbrainz.android.model.Review
-import org.listenbrainz.android.model.SearchResult
-import org.listenbrainz.android.model.SimilarUserData
-import org.listenbrainz.android.model.SocialData
-import org.listenbrainz.android.model.SocialResponse
-import org.listenbrainz.android.model.feed.FeedEvent
-import org.listenbrainz.android.service.SocialService
+import org.listenbrainz.shared.model.Review
+import org.listenbrainz.shared.model.SearchResult
+import org.listenbrainz.shared.model.SimilarUserData
+import org.listenbrainz.shared.model.SocialData
+import org.listenbrainz.shared.model.SocialResponse
+import org.listenbrainz.shared.model.feed.FeedEvent
+import org.listenbrainz.shared.service.SocialService
 import org.listenbrainz.shared.util.Resource
 import org.listenbrainz.shared.repository.AppPreferences
 import org.listenbrainz.shared.util.Utils.parseResponse
-
 
 
 class SocialRepositoryImpl(
