+++ sharedTest/src/main/java/org/listenbrainz/sharedtest/mocks/MockSocialRepository.kt
@@ -1,15 +1,15 @@
 package org.listenbrainz.sharedtest.mocks
 
-import org.listenbrainz.android.model.PinData
-import org.listenbrainz.android.model.RecommendationData
+import org.listenbrainz.shared.model.PinData
+import org.listenbrainz.shared.model.RecommendationData
 import org.listenbrainz.shared.model.ResponseError
-import org.listenbrainz.android.model.Review
-import org.listenbrainz.android.model.SearchResult
-import org.listenbrainz.android.model.SimilarUserData
-import org.listenbrainz.android.model.SocialData
-import org.listenbrainz.android.model.SocialResponse
-import org.listenbrainz.android.model.feed.FeedEvent
-import org.listenbrainz.android.repository.social.SocialRepository
+import org.listenbrainz.shared.model.Review
+import org.listenbrainz.shared.model.SearchResult
+import org.listenbrainz.shared.model.SimilarUserData
+import org.listenbrainz.shared.model.SocialData
+import org.listenbrainz.shared.model.SocialResponse
+import org.listenbrainz.shared.model.feed.FeedEvent
+import org.listenbrainz.shared.repository.social.SocialRepository
 import org.listenbrainz.shared.util.Resource
 import org.listenbrainz.sharedtest.testdata.SocialRepositoryTestData.testFollowersSuccessData
 import org.listenbrainz.sharedtest.testdata.SocialRepositoryTestData.testFollowingSuccessData
